Fixed rating parsing
Was using parseInt before we changed to 0.1 increments. Now we have to use parseFloat so it doesn't round.

+++ src/js/browse-movies.js
@@ -195,14 +195,14 @@ function get_rating_between_filter() {
 
 	if (below.checked) {
 		rating[1] = document.querySelector("#search #filters #rating_filters #below_rating").value;
-		rating[1] = parseInt(rating[1]);
+		rating[1] = parseFloat(rating[1]);
 	} else if (above.checked) {
 		rating[0] = document.querySelector("#search #filters #rating_filters #above_rating").value;
-		rating[0] = parseInt(rating[0]);
+		rating[0] = parseFloat(rating[0]);
 	} else if (between.checked) {
 		rating[0] = document.querySelector("#search #filters #rating_filters #between_start").value;
 		rating[1] = document.querySelector("#search #filters #rating_filters #between_end").value;
-		rating = [parseInt(rating[0]), parseInt(rating[1])];
+		rating = [parseFloat(rating[0]), parseInt(rating[1])];
 	}
 
 	if (rating[0] === "") {
